+++ conference-java/History.md
@@ -1,5 +1,10 @@
 # History of implementation
 
+### 2 - Spring Configuration using Java
+* AppConfig
+* Get rid of hardcoded instantiation
+* Use application context in main method
+
 ### 1 - Add Model, Repository and Service for Speaker (without Spring)
 * Speaker
 * SpeakerRepository